Update spring boot version and azure spring boot version
