+++ set-env-variables-template.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
-export COSMOSDB_URI=<put-your-COSMOS-DB-documentEndpoint-URI-here>
-export COSMOSDB_KEY=<put-your-COSMOS-DB-primaryMasterKey-here>
-export COSMOSDB_DBNAME=<put-your-COSMOS-DB-name-here>
+export COSMOS_URI=<put-your-COSMOS-DB-documentEndpoint-URI-here>
+export COSMOS_KEY=<put-your-COSMOS-DB-primaryMasterKey-here>
+export COSMOS_DATABASE=<put-your-COSMOS-DATABASE-name-here>
