+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
-azure.cosmosdb.uri=${COSMOSDB_URI}
-azure.cosmosdb.key=${COSMOSDB_KEY}
-azure.cosmosdb.database=${COSMOSDB_DBNAME}
+azure.cosmos.uri=${COSMOS_URI}
+azure.cosmos.key=${COSMOS_KEY}
+azure.cosmos.database=${COSMOS_DATABASE}
