--- src/test/resources/test.properties
@@ -1,1 +1,0 @@
-azure.keyvault.enabled=false
